[MOB-15749] -- Move CheckboxListView into shared-core-ui
