+++ plugins/hh-carnival/build.gradle.kts
@@ -11,6 +11,7 @@ repositories {
 
 dependencies {
     implementation(project(":shared-core-utils"))
+    implementation(project(":shared-core-ui"))
     implementation(project(":shared-core-freemarker"))
 
     implementation(kotlin("stdlib-jdk8"))
