+++ shared/core/ui/build.gradle.kts
@@ -10,5 +10,7 @@ repositories {
 }
 
 dependencies {
+    implementation(project(":shared-core-utils"))
+
     implementation(kotlin("stdlib-jdk8"))
 }
