+++ shared/core/ui/src/main/kotlin/ru/hh/plugins/models/CheckBoxListViewItem.kt
@@ -1,4 +1,4 @@
-package ru.hh.android.plugin.core.ui.model
+package ru.hh.plugins.models
 
 
 interface CheckBoxListViewItem {
